touch_monitor: Interprets EVIOCGMTSLOTS result correctly.
EVIOCGMTSLOTS, which is defined in <uapi/linux/input.h>, returns an
array of 32-bit signed integers. Previously we used array.array('l') to
convert the returned binary data to Python lists. This might not work on
some platforms since integer sizes in Python array module is
platform-dependent. We should replace it with struct.unpack to ensure
the result is always treated as an array of 32-bit signed integers.

BUG=none
TEST=manually on Caroline.

Change-Id: I6f378243fa179135bc182c7e37806225fb50837f
Reviewed-on: https://chromium-review.googlesource.com/523284
Commit-Ready: Youcheng Syu <youcheng@chromium.org>
Tested-by: Youcheng Syu <youcheng@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>