tools: Make 'add_files' in manifest.yaml optional for finalize_bundle.
'add_files' seems to be optional (there is already code checking if
add_files is included in manifest or not), especially for cros_payload
based bundle now.

However, currently there is a for-loop directly tries to fetch
`add_files` and iterate from that, which makes a KeyError if `add_files`
is not defined.

Changed to return empty list so add_files can really be optional.

BUG=None
TEST=finalize_bundle MANIFEST.yaml # without add_files

Change-Id: Ib0e9c5e4fe7c8982022280bdd08a407d4aebc56b
Reviewed-on: https://chromium-review.googlesource.com/526912
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Youcheng Syu <youcheng@google.com>