wireless_radiotap: Fix radiotap fields.
There's a field with bit 22 defined in
https://chromium.googlesource.com/chromiumos/third_party/kernel/+/4faa487cc797259f966c07d7d4327d8484d9a598
Define it in wireless_radiotap pytest, so the packet parsing would be
correct.

Also add some None for unused fields, so the bit number for MCS, AMPDU
status and VHT is correct.

Also add check that unused fields are not used, and log a warning
otherwise.

BUG=None
TEST=manually on DUT, check that wireless_radiotap can be passed, and
     got correct signal value for main and aux.

Change-Id: Idfb7ec7d800526acecd09c182ec9cb33b1d24650
Reviewed-on: https://chromium-review.googlesource.com/530584
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>