tools: finalize_bundle: Change working directory to directory of manifest.
Previously finalize_bundle takes input arg as bundle directory to work
on, and loads MANIFEST.yaml from it.

When the arg changed to manifest file, it also changed the default
bundle directory to cwd, and this breaks old workflow that people used
to extract bundles some where, and give that path to finalize_bundle.

It is possible to override that with --dir, but to simplify the
migration, it seems more reasonable to just change --dir default to
dirname(yaml).

BUG=None
TEST=./finalize_bundle PATH_TO/manifest.yaml

Change-Id: Iebcdef6d814f819eecffb5728bfc1991c159270a
Reviewed-on: https://chromium-review.googlesource.com/535093
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Youcheng Syu