cutoff: Change Finalize and FinalizeFQC to Shopfloor Service API NotifyEvent.
The Finalize and FinalizeFQC functions should be replaced by new
NotifyEvent, with event code as Finalize and Refinalize.

In Shopfloor Service API, the first parameter can be a
FactoryDeviceData; and in first implementation we only include serial
number.

Also revised cutoff and inform scripts to use curl and display TTY in
a way easier to debug.

BUG=chromium:640162
TEST=SERIAL_NUMBER=12345 ./generate_finalize_request factory_reset > x

Change-Id: Iea2a01b611a30a6d9f032cd8fa26d1efd0d63306
Reviewed-on: https://chromium-review.googlesource.com/538734
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>