raiden_charge: Add delay after device engaging.
Before doing the TestCharging, this test use Check5VInaCurrent to detect
the current is ready for the following charging test, and we found that
we usually failed here due to the unstable current. So, We add a delay to
make the charging current more stable.

TEST=none
BUG=none

Change-Id: I19e35496991ef9ddf043d4807db503b5159fffcb
Reviewed-on: https://chromium-review.googlesource.com/538499
Reviewed-by: Wei-Han Chen <stimim@chromium.org>
Tested-by: Yilun Lin <yllin@chromium.org>
Commit-Queue: Wei-Han Chen <stimim@chromium.org>