probe: match: add number comparison
For some components, we want to check if certain value is in a limit.
For example, "charge_full_design" of battery component is a number.  And
we should check if this number is greater than a number.  (Sometimes,
two battery that are considered the same components might have different
designed capacity, and we don't really care as long as it's large
enough).

BUG=None
TEST=make test

Change-Id: I07afe1151d13f0859213df7e125eef1943ed989e
Reviewed-on: https://chromium-review.googlesource.com/542675
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Yong Hong <yhong@chromium.org>