factory: add new test status "SKIPPED"
In factory, we have two similar but different concepts, "skip" and
"waive".  A test item is "skipped" if it should not be run.  A test item
is "waived" if it should be run, but the failure will be ignored.

In old implementation, "FAILED_AND_WAIVED" is a test status, while
"SKIPPED" was not.  A "SKIPPED" test was marked as "PASSED" with a
special tag.  This difference makes developers confuse, because they
need to deal with two different logic.  And we also need additional code
to check if a test is skipped.

In this CL, new test status "SKIPPED" is added, so we can know if the
test is skipped by just look into its status.

BUG=b:62894380
TEST=make test

Change-Id: Ia1be0381be129379edc7e90c94eab0462a28440a
Reviewed-on: https://chromium-review.googlesource.com/544443
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>