pytests: Change hwid.py to symlink of hwid_v3.
The 'hwid.py' was actually an implementation for  'HWID v2' test item
which has been deprecated for a long time since we don't have any
devices using HWID v2 database now and needs more factory or
RMA processing.

To simplify that, we are changing 'hwid.py' to be a link to "current
(recommended) HWID test - which implies v3. And in future when there
is v4 or newer versions, we may change the symlink to that.

BUG=None
TEST=make test

Change-Id: I2b96358e35acda5db01f227b0d256da2d9383028
Reviewed-on: https://chromium-review.googlesource.com/545378
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Yong Hong <yhong@chromium.org>