net_utils: SetEthernetIp: allow netmask be int
`net_utils.GetEthernetIp(interface, True)` will return a tuple of
(ip, netmask), where netmask is an integer (number of prefix bits).
However, if you use the return value on SetEthernetIp directly:
`net_utils.SetEthernetIp(ip, interface, netmask)`
You will get an exception because it failed to construct a ifconfig
comamnd for you.  `SetEthernetIp` excepts netmask to be a string like
'255.255.255.0'.

In this CL, when argument netmask is an integer, we will convert it into
a string, so the function will work properly.

BUG=None
TEST=manual

Change-Id: I8f32204d14f2e25553a544f8f3575b60ee21c59c
Reviewed-on: https://chromium-review.googlesource.com/547164
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>