umpire: Abort importing resource if version can't be identified.
Dome expects 'version' to be available in all Umpire payloads JSON
files, however Umpire and cros_payload both don't fail if version is
missing or empty.

It should be either cros_payload or Umpire to abort invalid resource.
Currently the version is detected and written into JSON config by
cros_payload, but in future if we want to make cros_payload more
generic, it will be the caller of cros_payload to find out resource
property and set by 'cros_payload add_meta' command.

As a result, since Umpire is the one knowing "resources need version for
Dome to use" and the one that really reads version (for toolkit to
update), we want Umpire to check and abort if version is empty or
missing.

BUG=chromium:737465
TEST=manually tested. cros_docker.sh umpire test

Change-Id: I763f48cde7bce908e928363b3d6d5f4a7be1ab90
Reviewed-on: https://chromium-review.googlesource.com/552420
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Youcheng Syu <youcheng@chromium.org>