test: Move device data functions to a new module 'device_data'.
The 'device data' is provided by state shelves, but it's used more often
and need to be shared between Goofy, tests, and shopfloor services. Also
there are lots of constant names dedicated to device data.

So we are moving all the key and constant names, logic and helper
functions into a new module "device_data" so it is more clear for how to
use device data.

BUG=chromium:640162,b:38154177
TEST=make test

Change-Id: I462d3ba7148eb83f9eb2579760a8c5d6aa1ac6f1
Reviewed-on: https://chromium-review.googlesource.com/559187
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>