Instalog: Change the lifespan of temporary directory in plugins
Change the lifespan of temporary directory in plugins.
Before this change, the plugin's temporary directory is created in SetUp
and deleted in TearDown. However, if the temporary directory disappears
unexpectly, the plugin will always fail in the loop. After this change,
the plugin will fail once and try again with a new temporary directory.

BUG=b:36667063
TEST=make test

Change-Id: I41895b15dbfe2c0a6011fabbd1965ee09e6cff60
Reviewed-on: https://chromium-review.googlesource.com/559208
Reviewed-by: Joel Kitching <kitching@chromium.org>
Tested-by: Chun-Tsen Kuo <chuntsen@chromium.org>
Commit-Queue: Chun-Tsen Kuo <chuntsen@chromium.org>