pytests: Add new 'exec_shell' to replace line_check_items.
The 'line_check_item' was powerful but is also hard to use when there
are only few commands to invoke, especially that it lacks of the ability
to watch progress and output in real time.

The new 'exec_shell' is dedicated to invoking commands, with better
output in real time and easier to setup.

BUG=None
TEST=make test; manually verified exec_shell test.

Change-Id: I1085526b6f87a36447a933edf12022dd624c6ca2
Reviewed-on: https://chromium-review.googlesource.com/560928
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>