Dome: set Cache-Control: no-cache in HTTP response header
By default, nginx does not send Cache-Control header to the client, and
Chrome would use cache instead of revalidating the resource in a short
time period. This would cause trouble for developers and even users
(their code may not be synced and so may not work). A better approach is
to tell Chrome to always revalidate. But the server is free to respond
with a status code 304 (Not Modified) to avoid sending the file again.

BUG=b:35305024
TEST=Manually tested:
    1. Open Chrome and its developer console
    2. Go to settings, uncheck "Disable cache (while DevTools is open)"
    3. Go to the network tab
    4. Navigate to URL http://localhost:8000/
    5. Make sure "Cache-Control: no-cache" is in each response header

Change-Id: I8c397810ebd62ea005f98beeb6f29428b80675e3
Reviewed-on: https://chromium-review.googlesource.com/569526
Reviewed-by: Youcheng Syu <youcheng@chromium.org>
Commit-Queue: Mao Huang <littlecvr@chromium.org>
Tested-by: Mao Huang <littlecvr@chromium.org>