Aquarium - A button to toggle full screen
In run-in test, a webgl-based aquarium is used to test the
graphics. In a a long-running run-in test, the static parts
of our goofy UI might cause a burn-in on the LCD panel.

A button is added to the aquairum iframe, which can be used
to toggle full screen mode.

By default, it goes to full screen mode when the tests starts.

BUG=chromium:728186
TEST=manually

Change-Id: I52de045a7fd88473b73d1fc89cc5f8e3fd39a6af
Reviewed-on: https://chromium-review.googlesource.com/569941
Commit-Ready: Shen-En Shih <petershih@chromium.org>
Tested-by: Shen-En Shih <petershih@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>