Change imports in testlog to relative import syntax.
Change imports in testlog from using "import xxx" to "from . import
xxx", so it's clear that it's a relative import.

Although relative imports are forbidden in Chromium OS Python style
guide, we still want the benefit that the testlog module to be
independent to the factory module, so we still want to use relative
import here.

BUG=None
TEST=make test

Change-Id: Ica7b376684e0cc4d7797cc1ee7f753bdbe41fb64
Reviewed-on: https://chromium-review.googlesource.com/571248
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Chun-ta Lin <itspeter@chromium.org>