goofy: Fix "console.log does not exist" error.
In the current implementation, the /var/factory/log/console.log is
delay-created until the first log message is generated. On the other
side, a daemon thread is spawned in goofy initialization phase,
which 'tail -F' the console.log, and send it out to our UI interface.
Combine these two facts, the tail process generates an error message
in the first run, saying console.log does not exist.

The reason why the file 'console.log' is delay-created, is because
the file handler is created at module-loading time, and it is
expected that the console log file is not created at that time.

In this commit, the console.log is created (if not created yet)
before tailing it.

BUG=chromium:683060
TEST=rm /var/factory/log/console.log, restart goofy, check
/var/log/factory.log should have no tail error, and check a text
"Opened console" should be printed on the goofy web interface.

Change-Id: Iad55f3914f3ee1ef11f1557993137eabca653a0e
Reviewed-on: https://chromium-review.googlesource.com/573886
Commit-Ready: Shen-En Shih <petershih@chromium.org>
Tested-by: Shen-En Shih <petershih@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>