config_utils: Change inheritance to use C3 linearization
Change from simple depth-first search based inheritance to C3 superclass
linearization algorithm.

The original algorithm is same as old-style class inheritance in Python,
and C3 superclass linearization is same as new-style class inheritance
in Python.

This solves the problem that the following config has final value of b
as 0, which is unintuitive and cause problem when used the config for
JSON test list:

-- base.json --
{"b": 0}
-- a.json --
{"inherit": ["base"]}
-- b.json --
{"inherit": ["base"], "b": 42}
-- final.json --
{"inherit": ["a", "b"]}

BUG=chromium:738969
TEST=make test

Change-Id: I13619593cdebe45afb72be233c52f92de9c26a66
Reviewed-on: https://chromium-review.googlesource.com/575789
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>