goofy: Fix error 'thermal service is not found'
Two services are tried to gain more control on cpu frequency.
One of them is 'thermal' service, the other one is 'dptf'.
It cannot be easily determined which one should be used, so
in current design, we tried to control both of them, and
typically one of them will failed, generating an error message
like 'thermal service is not found' in the factory.log.

At least three methods can be used to check if a given service
name exists or not:
1. Check files under /etc/init folder.
2. Check return code of 'status' command
   It returns 0 if the service exists; otherwise returns 1.
3. Create a unix socket to /com/ubuntu/upstart, and doing some
   communication with it.

In this commit, the second method is chosen. A new method is
added to the service utils, which checks if a given service
name exists or not.

BUG=chromium:683060
TEST=run factory toolkit, there should have no error message
like 'thermal service is not found' in factory.log

Change-Id: Ic295ac275b617fe3e03090b65c9afcccc231be44
Reviewed-on: https://chromium-review.googlesource.com/575117
Commit-Ready: Shen-En Shih <petershih@chromium.org>
Tested-by: Shen-En Shih <petershih@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>