dome: add README.md in backend/migrations/ folder
This files reminds the developers that files in this folder should NOT
be modified in any circumstances. To change the database schema, the
developer must generate a new migration file. The README also gives a
brief instruction on how to update the database schema.

BUG=None
TEST=None

Change-Id: Ib4ae52aecf50ca34c84b1b4b529e17c236b3eaf8
Reviewed-on: https://chromium-review.googlesource.com/580015
Commit-Ready: Mao Huang <littlecvr@chromium.org>
Tested-by: Mao Huang <littlecvr@chromium.org>
Reviewed-by: Yong Hong <yhong@chromium.org>