goofy - fix bug when multiple bluetooth devices in host
In a station-based bluetooth test, we ask DUT to connect to
the host to verify its bluetooth functionalities. However,
if there are multiple bluetooth devices exist in host, we
need to fix the interface name beforehand (e.g., 'hci0' or
'hci1'). But, in real world, an external bluetooth dongle
might sometimes appeared as 'hci0', and sometimes as 'hci1'.

In this commit, we support a new testing option, which uses
all the available bluetooth devices on the host, and asks
a DUT to connect to any one of them. In this way, we no
longer need to specify the host interface name beforehand.

BUG=none
TEST=manually

Change-Id: I8b7f05451643bb34518f38545cbe3db08f357b2d
Reviewed-on: https://chromium-review.googlesource.com/581048
Commit-Ready: Shen-En Shih <petershih@chromium.org>
Tested-by: Shen-En Shih <petershih@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>