test list: manager: move EvaluateRunIf to manager
We'd like to use the similar way to evaluate run_if and args.  Move the
implementation into manager.

Now run_if will be passed to 'eval' function with 'device' and
'constants' in the namespace.  So you have to use 'not' rather than '!'
now.  And 'and', 'or' or other more complicated syntaxes are supported
now.

BUG=b:31678190
TEST=manual

Change-Id: I4d10f8e2f0b94afe46e6cc3107de2a46166a50d9
Reviewed-on: https://chromium-review.googlesource.com/587700
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>