device: Support reading vpd from sysfs driver
For modern Chromebooks, the firmware will prepare a cache of
VPD data and exported it by kernel driver via sysfs.

The purpose of this CL, is to use the information under
/sys/firmware/vpd to speed up reading VPD values if the caller
does not need "live" values. For example, for goofy's startup
hooks.

However, it must be noted that, the data there might not be updated
when new values are written to the VPD. This renders most of the
tests in factory software inapplicable.

In this sense, the inherit graph of VPD components are re-designed,
and a new 'boot' property is added to vpd. From this 'boot' property,
one can get the VPD values cached at boot time very quickly. An
example usage is for a board hook. It can use this to read VPD data
in the init hook, reducing goofy's startup time.

BUG=chromium:683060
TEST=make test
TEST=manually test

Change-Id: I87c27e915d787914d2caaea0bb1d5bfdb6d2c565
Reviewed-on: https://chromium-review.googlesource.com/593711
Commit-Ready: Shen-En Shih <petershih@chromium.org>
Tested-by: Shen-En Shih <petershih@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>