factory: add FlattenGroup
Sometimes we want to define a group of tests but don't want to put them
in an additional layer.

For example, in Components test group, we put a small stress test +
countdown test.  Without FlatttenGroup, the test list structure will be:

  Components
  |-- Wifi
  |-- Probing
  |-- Bluetooth
  `-- StressGroup
      |-- Countdown
      `-- StressAppTest

This prevents the Components group running in parallel.

With FlattenGroup, you can do the following:

  "StressGroup": {
    "inherit": "FlattenGroup",
    "subtests": [
      "Countdown",
      "StressAppTest"
    ]
  }

And the Components group will be resolved to:

  Components
  |-- Wifi
  |-- Probing
  |-- Bluetooth
  |-- Countdown
  `-- StressAppTest

Also, allowing a FlattenGroup can reduce the number of layers shown on
UI, makes the UI more clean.

BUG=b:31678190
TEST=make test, manual

Change-Id: I1292f0655cdcfbdd5fdc16961fa8331e710bcfd8
Reviewed-on: https://chromium-review.googlesource.com/593487
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>