goofy: Update system info in an async way during init
The system info is now initialized asynchronously by
posting an UPDATE_SYSTEM_INFO event to the event server.

This speeds up the start up time greatly, but now the
'system_info' in the shared data might be empty for a
short time window. All callers are checked to ensure
the correctness. The only one needs to be modified is
the status monitor plugin, and it's now also update from
'system_info' periodically in a previous commit.

BUG=chromium:683060
TEST=make test
TEST=restart goofy, check status monitor can show correct
system info on UI after a few seconds.

Change-Id: I2b0fd9d81e8595571367e472bbd0aefe3daedbd5
Reviewed-on: https://chromium-review.googlesource.com/593712
Commit-Ready: Shen-En Shih <petershih@chromium.org>
Tested-by: Shen-En Shih <petershih@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>