status-monitor: Fetch device info from `device.info`
When the device info is changed, the 'UpdateDevieInfo' RPC
call will be called. The status monitor plugin should
update the device info on UI after received this call.

To support this, I choose to modify the 'GetSystemStatus'
RPC call. Originally, this call only returns the dynamic
system status like CPU loading and battery info. Now,
it also returns the updated device info, if the device info
is needed to be updated.

To reflect this change, the name of this RPC call is also
changed from 'GetSystemStatus' to 'GetSystemInfo'.

BUG=chromium:751026
TEST=make test

Change-Id: I4091a9482fc0e99abbf5459283ac26ea5c958ac0
Reviewed-on: https://chromium-review.googlesource.com/595790
Commit-Ready: Shen-En Shih <petershih@chromium.org>
Tested-by: Shen-En Shih <petershih@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>