test_list_iterator: don't clear passed automated sequence
Tests in AutomatedSequence depends on each other, so if we are going to
retest any of them, we need to retest all of them.  However, current
implementation has a bug that no matter what current status of the group
is, the test status will be cleared.

Since a test group is marked as PASSED iff all subtests are PASSED, if
the automated sequence is PASSED, we don't want to reset it.  (An
exception is, status_filter is not given or PASSED is in status_filter).

BUG=b:64178558
TEST=make test, manual

Change-Id: Ibfb3b2712e130ce8f28da1ba958157a318c25a41
Reviewed-on: https://chromium-review.googlesource.com/595887
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Tested-by: Chia-Hsiu Chang <chia-hsiu.chang@quantatw.com>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Chia-Hsiu Chang <chia-hsiu.chang@quantatw.com>