blocking_charge: Add options to get target from cutoff / goofy.
Since the min_charge_pct is not in test list options anymore, we can't
pass it in from test list as before. Add special values to the
pytest to indicate that the target level should be either read from
config of cutoff (cutoff.json), or be read from config of Goofy plugin
charge_manager, so we don't need to write the level twice.

Also improve the document.

BUG=chromium:737920
TEST=make test
TEST=manually on DUT.

Change-Id: Ie7c9ef57187a0421b0ca50ea1d94777d9fe310c6
Reviewed-on: https://chromium-review.googlesource.com/597562
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>