manager: prevent infinite recursive call
TestList implements __getattr__, which will be triggered when an
attribute cannot be found in normal lookup process.  However, if a
property attribute (@property) raises an AttributeError exception,
__getattr__ will be triggered too.  And this might cause infinite
recursive call (because __getattr__ relies on ToFactoryTestList(), which
might rely on property attribute).

Because ToFactoryTestList should never call itself (directly or
indirectly).  A @NoRecursion decorator is added to prevent this.  So the
traceback will be shorter.

We will add another wrapper to prevent property attribute raising
AttributeError in next CL.

BUG=b:31678190
TEST=make test

Change-Id: I9ff64620524d849ced1f844a152424f5c30965d9
Reviewed-on: https://chromium-review.googlesource.com/598753
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>