test list: add skipped_tests and waived_tests
For early builds, some test items need to be waived or skipped.  This is
currently implemented by <board>_hooks.py in each board overlay (most of
them just copy and paste from previous projects).

This CL standardizes how to mark a test as skipped or waived.
'skipped_tests' and 'waived_tests' are added to test list options.  Both
of them should have the format:

  {
    <PHASE>: [ <test path>... ]
  }

For example:

  {
    "PROTO": [
      "SMT.AudioJack",
      "SMT.SpeakerDMic"
    ],
    "EVT": [
      "SMT.AudioJack"
    ]
  }

We force you to specify the phase, so when a new build starts, the
default skipped list and waived list become empty.  TPM or SWE have to
add tests to the list explicitly.

Also, to force the new approach to be used, the 'waived' is not
supported by test_object.  So the following test object:

  {
    "inherit": "LED",
    "waived": true
  }

will cause an exception.

"test_list.SetSkippedAndWaivedTests()" is called by Goofy, only for the
active test list.  Because marking a test as SKIPPED needs to interact
with state instance, and `test_list.state_instance` is set by Goofy.
(And we don't care about what test is skipped or waived for non-active
test lists.)

BUG=b:31678190
TEST=make test

Change-Id: Ie471093c225bccba345a255e233c4884a065ba14
Reviewed-on: https://chromium-review.googlesource.com/599928
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>