test list: allow pattern matching for skipped tests
If a test is skipped in SMT, it is usually skipped in other stations
too.  With this CL, instead of writing:

    [
        'SMT.AudioJack',
        'FFT.AudioJack',
        'FAT.AudioJack',
        'FOO.BAR.AudioJack'
    ]

Now you only need to write:

    [ '*.AudioJack' ]

to skip all AudioJack tests.

Note that 'SMT.AudioJack-2' is not matched.  (Because this might be a
test for different audio component, and should not be skipped).

BUG=b:31678190
TEST=make test

Change-Id: I48d74e415825e11e5fcf1ad2ba932fd2442d57f9
Reviewed-on: https://chromium-review.googlesource.com/600347
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>