test list manager: prevent reloading invalid test list
If a test list is invalid, ReloadIfModified will keep current test list
config unchanged.  But the timestamp of test list config is not changed
either, so it will try to reload the same invalid test list again and
again.  This is very bad (we have to construct the entire test list and
rollback everytime).

In this CL, we update the timestamp of test list config to current time
if we failed to load it.

BUG=b:31678190
TEST=make test

Change-Id: I215def818cfc3635706d404a313f67e3386c6053
Reviewed-on: https://chromium-review.googlesource.com/601775
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>