goofy: add note when test list modification is detected
When we try to reload the test list, no matter it success or not, a note
will be added to Goofy by calling AddNote RPC call.  If we success, log
level will be INFO, if we fail, log level will be WARNING.

Also change Goofy UI, if the latest note is a warning, updateNote() will
show a popup window.

BUG=b:31678190
TEST=make test, manual

Change-Id: I9ef6b93652520c7edf57c594c6d152ce02e40c91
Reviewed-on: https://chromium-review.googlesource.com/602408
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>