goofy: Do not wait socket in launch_chrome()
A timeout of 90 seconds is located in launch_chrome() to
wait a web socket to connect. This timeout was added in an
aged CL to try to reboot Chrome browser several times,
since back then the Chrome browser seems not stable enough.

As time goes, the Chrome browser is getting stable, and the
retrying loop has already been removed. Furthermore, the
launch_chrome() can now be removed.

Noted that, the caller still waits for at least one web socket
to get connected, so this does not modify the flow much. Also
noted that, we can use init script to set up the Chrome startup
webpage if needed.

BUG=chromium:734925
TEST=make test
TEST=manually test

Change-Id: I33c03206a147ee8e5aa3a78162b2cd2657cfc25b
Reviewed-on: https://chromium-review.googlesource.com/603587
Commit-Ready: Shen-En Shih <petershih@chromium.org>
Tested-by: Shen-En Shih <petershih@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>