goofy: read device data from vpd in init_states
To make sure that device data is synced with VPD data, we often add
'read_device_data_from_vpd' to the beginning of each station.  Move this
to goofy.init_states so we don't need to do it explicitly anymore.
(This shouldn't break current flow, because 'write_device_data_to_vpd'
should be called as soon as the device_data is updated).

Also, having this change allows us to use device data in
`options.skipped_tests['<expr>']`.  So the logic in
<board>_hooks.py to "skip SMT tests if factory.end_SMT is set" is no
longer required, the new approach is:

    with TestList(...) as test_list:
      test_list.options.skipped_tests = {
          'device.factory.end_SMT': ['SMT']
          'device.factory.end_FRT': ['FRT']
          'device.factory.end_RunIn': ['RunIn']
      }
    # or equivalent code in JSON test list.

BUG=b:31678190
TEST=make test

Change-Id: I7b279233b83d45d71fc51ace15166c4c9e699787
Reviewed-on: https://chromium-review.googlesource.com/605418
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>