goofy: Silent the error when event server is stopped.
We use net_utils.ShutdownTCPServer to stop event server, which connect
to the server and close immediately, and would cause a broken pipe when
event server sends the hello message. Handle broken pipe the same way as
the client quitting, and ignore the exception.

BUG=None
TEST=make test
TEST=manually on DUT, run factory_restart and see no broken pipe error
     in factory.log

Change-Id: I58c5dcd535fc6e6d68a1f0c0bbc105c203a332aa
Reviewed-on: https://chromium-review.googlesource.com/611881
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>