test list manager: remember mtime for all dependencies
We used to compare latest mtime of test lists on filesystem with the
timestamp of TestListConfig object.  In this case, we will fail to
detect file modification if some files have timestamp in future.

Let's remember timestamps of all config files, and if any of them
changes (newer, older, or disappear), we count it as modified.

BUG=b:31678190
TEST=make test

Change-Id: I6c4f86a51965a8fea5829bc3ac2f8b557b5caf92
Reviewed-on: https://chromium-review.googlesource.com/613040
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>