Umpire: revise Umpire services config to support JSON schema
To add an online editor of Umpire services on Dome, we need to send the
services config schema to front-end. However, since the front-end was
implemented with javascript, we need to migrate the schema from
utils/schema to JSON format.

This change:
1. Adds a new class "JSONSchemaDict" in utils/schema.py, allowing
nesting JSONschema in factory's schema.
2. Moves service config schema from CONFIG_SCHEMA in <service_name>.py
to <service_name>_config.schema.json
3. Rewrites the "GetServiceSchemata()" function in umpire_service.py.
4. Adds a new function "GetAllServiceSchemata()" in umpire_service.py.

BUG=b:64462600
TEST=manually

Change-Id: I4523deda29de91b0065de43d4aa83130150b1461
Reviewed-on: https://chromium-review.googlesource.com/615341
Commit-Ready: Chih-Wei Ning <cwning@google.com>
Tested-by: Chih-Wei Ning <cwning@google.com>
Reviewed-by: Youcheng Syu <youcheng@chromium.org>