factory_server: Support JSON-RPC and WSDL/SOAP web services.
The Chrome OS Factory Shopfloor Service was defined to be using XML-RPC
protocol. However, partners using Windows servers usually prefer to
implement web services using VS.NET, which implies SOAP/WSDL.

This change adds wrappers so the manufacturing line may deploy JSON-RPC
or WSDL/SOAP web service as shopfloor service.

To use JSON-RPC, prefix a 'jsonrpc:' to the service URL.
To use WSDL, prefix a 'wsdl:' to the service URL.
WSDL is automatically used if the lower cased URL ends with 'wsdl'.

Additionally, describing the generic compound type in WSDL may be not
trivial in certain framework (also that we need 'dot' in key name,
although this is allowed in SOAP). An additional filter 'json:' is
created that it'll translate all function call arguments and return
value as JSON string, so WSDL can simply declare the parameters as
'string'. This is also helpful for XML-RPC implementations that do not
support 'nil' extension.

Examples:

  http://10.3.0.11  # A XML-RPC server.
  json:http://10.3.0.11  # A XML-RPC server with input/output in JSON.
  jsonrpc:http://10.3.0.11  # A JSON-RPC server.
  http://10.3.0.11/?wsdl  # A WSDL (SOAP) server.
  json:http://10.3.0.11/?wsdl  # A WSDL server with input/output in JSON.

BUG=chromium:755995
TEST=make test; cros_docker.sh umpire test
     Also tried calling GetDeviceInfo via a mock shopfloor service.

Change-Id: I33cc12692cb614f6a1243c2868d16daf06de36b4
Reviewed-on: https://chromium-review.googlesource.com/616940
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>