goofy.js: Bind the test visible state with tab.
Bind the test visible state with tab, so tab selection actually changes
test visible state.

Also stop setting visible when select item on the tree view, since we
have tab now, and it would cause focus lost for the context menu. An
item to bring the UI to front would be added to the context menu in a
later CL.

BUG=None
TEST=make test
TEST=manually on DUT.

Change-Id: I3507022ba8e40a3237fb1fd9c12dd95c2b7a2a4c
Reviewed-on: https://chromium-review.googlesource.com/622110
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>