goofy.js: Minimize tab flicker when test is finished.
When a test is finished and the tab is removed, tabbar would
automatically select the next tab available, and the goofy backend would
set visible on the first active test, which cause a small flicker that
shows a wrong tab selected.

Fix this by selecting the first tab in front end after a test is
finished.

BUG=b:64963909
TEST=make test
TEST=manually on DUT.

Change-Id: I406d16015475896b47aeb76f286b6479a320e186
Reviewed-on: https://chromium-review.googlesource.com/625541
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>