goofy: Preserve tab ordering in parallel test group.
Currently for parallel test group with multiple UI, the order for which
the tab appearing is not deterministic, since it's the order that those
tests call test_ui.UI(). Send a INIT_TEST_UI before starting invocation
if the test have UI, to make sure that the test UI are initialized in
order, so it would be same as the order in test list.

Also add some comment on why we still need INIT_TEST_UI in test_ui.UI().

BUG=b:64963909
TEST=make test
TEST=manually on DUT.

Change-Id: Ia90c12eec4b2875d6d852fa79db52761abb0a12b
Reviewed-on: https://chromium-review.googlesource.com/625542
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>