pytests: Remove shopfloor task from 'start' test.
The 'start' used to support checking a scanned (or read from VPD) serial
number and check if it is valid using shopfloor. This actually won't
work because
 (1) Most shopfloor backends today do not support CheckSerialNumber API.
 (2) Most projects need to start with only MLB SN.
 (3) The reference test doesn't call 'start' anymore.

As a result, we move the 'No server URL' warning to sync_factory_server
and remove the shopfloor / VPD-SN tasks from start.

BUG=chromium:760017
TEST=make test

Change-Id: If9817a7c4b23a70fbc745282319d7f36684ac799
Reviewed-on: https://chromium-review.googlesource.com/641011
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>